First draft of SSX "Model A"
Model A uses parity bases (sin/cos) to have full S14 boundary
conditions on all walls. It's a square duct with Coulomb gauge vector
potential form for the induction equation.

The only simplifications from S14 are

* constant resistivity
* constant **kinematic** viscosity
* constant thermal conductivity

The latter two are somewhat non-sensical, since this is fully
compressible.

The main loop is not yet implemented, but the solver builds.

Initial conditions are provided by a new spheromak implementation that
follows S14 in specifying J0 and then solving a Vector Poisson's
equation to get A0. J0 is not yet implemented.